endpoints: catch only at top level, fail on errors
Our intermediate `catch{}` to log the error was making failures in seed
not result in HTTP 500s.
