+++ src/endpoints/seed/biographies.ts
@@ -14,10 +14,9 @@ export const seedStaff = async (
 
   const placeholder = await fetchFileByURL(
     'https://upload.wikimedia.org/wikipedia/commons/f/f8/Profile_photo_placeholder_square.svg',
-  ).catch((e) => payload.logger.error(e))
+  )
   if (!placeholder) {
-    payload.logger.error(`Downloading placeholder photo returned null...`)
-    return {}
+    throw new Error(`Downloading placeholder photo returned null...`)
   }
   const placeholders = await upsert('media', payload, incremental, tenantsById, (obj) => obj.alt, [
     ...Object.values(tenants).map(
@@ -38,10 +37,9 @@ export const seedStaff = async (
 
   const headshotData: { data: RequiredDataFromCollectionSlug<'media'>; file: File }[] = []
   for (const photo of headshots(tenants)) {
-    const image = await fetchFileByURL(photo.url).catch((e) => payload.logger.error(e))
+    const image = await fetchFileByURL(photo.url)
     if (!image) {
-      payload.logger.error(`Downloading ${photo.alt} photo returned null...`)
-      return {}
+      throw new Error(`Downloading ${photo.alt} photo returned null...`)
     }
     headshotData.push({
       data: {
